Moved `useDoubleClick` to new utils location
